The Request-Line parser, which reads the start of the file, then tokenize, then parse the HTTP's version, then returns

- the method string,
- the URI string,
- and the version major and minor numbers.

Because of this, only the version number text is validated.

Save Markiplier!

+++ src/httpParse.h
@@ -20,8 +20,8 @@
 struct http_request_line {
     char method[HTTP_METHOD_MAX_LENGTH + 1];
     char uri[HTTP_URI_MAX_LENGTH];
-    unsigned int major_version;
-    unsigned int minor_version;
+    unsigned long major_version;
+    unsigned long minor_version;
 };
 
 /* Parses the Request-Line in a string, checking the logic.
